moving aws specific code from sageworks.aws_service_broker to sageworks.core.cloud_platform.aws
