+++ aws_setup/aws_account_check.py
@@ -5,7 +5,7 @@ import logging
 import awswrangler as wr
 
 # SageWorks Imports
-from sageworks.aws_service_broker.aws_account_clamp import AWSAccountClamp
+from sageworks.core.cloud_platform.aws.aws_account_clamp import AWSAccountClamp
 from sageworks.utils.config_manager import ConfigManager, FatalConfigError
