+++ scripts/athena_ddl_mixed_case.py
@@ -11,7 +11,7 @@ will be lowercased to castle.
 
 
 # SageWorks Imports
-from sageworks.aws_service_broker.aws_account_clamp import AWSAccountClamp
+from sageworks.core.cloud_platform.aws.aws_account_clamp import AWSAccountClamp
 
 boto3_session = AWSAccountClamp().boto3_session
