+++ scripts/copy_data_catalog_db.py
@@ -1,4 +1,4 @@
-from sageworks.aws_service_broker.aws_account_clamp import AWSAccountClamp
+from sageworks.core.cloud_platform.aws.aws_account_clamp import AWSAccountClamp
 
 # Initialize a session using your AWS credentials
 aws_account_clamp = AWSAccountClamp()
