+++ src/sageworks/api/parameter_store.py
@@ -8,7 +8,7 @@ import base64
 from botocore.exceptions import ClientError
 
 # SageWorks Imports
-from sageworks.aws_service_broker.aws_session import AWSSession
+from sageworks.core.cloud_platform.aws.aws_session import AWSSession
 
 
 class ParameterStore:
