+++ src/sageworks/api/pipeline.py
@@ -9,7 +9,7 @@ import pandas as pd
 
 # SageWorks Imports
 from sageworks.utils.config_manager import ConfigManager
-from sageworks.aws_service_broker.aws_account_clamp import AWSAccountClamp
+from sageworks.core.cloud_platform.aws.aws_account_clamp import AWSAccountClamp
 from sageworks.core.pipelines.pipeline_executor import PipelineExecutor
