+++ src/sageworks/core/artifacts/artifact.py
@@ -7,7 +7,7 @@ import sys
 import logging
 
 # SageWorks Imports
-from sageworks.aws_service_broker.aws_account_clamp import AWSAccountClamp
+from sageworks.core.cloud_platform.aws.aws_account_clamp import AWSAccountClamp
 from sageworks.core.cloud_platform.aws.aws_meta import AWSMeta as Meta
 from sageworks.utils.sageworks_cache import SageWorksCache
 from sageworks.utils.aws_utils import sagemaker_delete_tag, dict_to_aws_tags
