+++ src/sageworks/core/transforms/transform.py
@@ -7,7 +7,7 @@ from typing import Union, final
 import logging
 
 # SageWorks Imports
-from sageworks.aws_service_broker.aws_account_clamp import AWSAccountClamp
+from sageworks.core.cloud_platform.aws.aws_account_clamp import AWSAccountClamp
 from sageworks.utils.config_manager import ConfigManager, FatalConfigError
