+++ src/sageworks/utils/ecs_info.py
@@ -2,7 +2,7 @@ import json
 import os
 
 # SageWorks Imports
-from sageworks.aws_service_broker.aws_account_clamp import AWSAccountClamp
+from sageworks.core.cloud_platform.aws.aws_account_clamp import AWSAccountClamp
 
 
 class ECSInfo:
