+++ src/sageworks/utils/endpoint_metrics.py
@@ -4,7 +4,7 @@ from datetime import datetime, timedelta, timezone
 import pandas as pd
 
 # SageWorks Imports
-from sageworks.aws_service_broker.aws_account_clamp import AWSAccountClamp
+from sageworks.core.cloud_platform.aws.aws_account_clamp import AWSAccountClamp
 
 
 class EndpointMetrics:
