+++ src/sageworks/utils/extract_model_artifact.py
@@ -12,7 +12,7 @@ import joblib
 import logging
 
 # SageWorks Imports
-from sageworks.aws_service_broker.aws_account_clamp import AWSAccountClamp
+from sageworks.core.cloud_platform.aws.aws_account_clamp import AWSAccountClamp
 
 log = logging.getLogger("sageworks")
