+++ src/sageworks/utils/sageworks_event_bridge.py
@@ -4,7 +4,7 @@ from pprint import pprint
 from botocore.exceptions import ClientError
 
 # Local Imports
-from sageworks.aws_service_broker.aws_account_clamp import AWSAccountClamp
+from sageworks.core.cloud_platform.aws.aws_account_clamp import AWSAccountClamp
 
 
 class SageWorksEventBridge:
