+++ src/sageworks/utils/sageworks_sqs.py
@@ -3,7 +3,7 @@
 import logging
 
 # Local Imports
-from sageworks.aws_service_broker.aws_account_clamp import AWSAccountClamp
+from sageworks.core.cloud_platform.aws.aws_account_clamp import AWSAccountClamp
 
 
 class SageWorksSQS:
