+++ tests/aws_account/aws_account_clamp_tests.py
@@ -1,7 +1,7 @@
 """Tests for the AWS Account Clamp"""
 
 # SageWorks Imports
-from sageworks.aws_service_broker.aws_account_clamp import AWSAccountClamp
+from sageworks.core.cloud_platform.aws.aws_account_clamp import AWSAccountClamp
 
 
 def test():
