+++ tests/create_test_artifacts/create_empty_model_group.py
@@ -6,7 +6,7 @@ Models:
 
 import logging
 from botocore.exceptions import ClientError
-from sageworks.aws_service_broker.aws_account_clamp import AWSAccountClamp
+from sageworks.core.cloud_platform.aws.aws_account_clamp import AWSAccountClamp
 
 # Setup the logger
 log = logging.getLogger("sageworks")
